+++ src/Type/Introspection.php
@@ -51,6 +51,7 @@ use GraphQL\Utils\Utils;
  * - typeIsOneOf
  *   Include field `isOneOf` for types?
  *   Default: false
+ *
  * @see \GraphQL\Tests\Type\IntrospectionTest
  */
 class Introspection